Finished most of the HTML/Django coding for the song module
This includes adding edit forms for the details and lyrics/chords.
Much of the views.py code was condensed to remove redundancy.
Lyrics/chord/details are now all handled by one view and another
for editing them. Still need to get the correct sort order for the
standard details view.